add / move test
